Seperated logic from data
